+++ src/main/java/org/torproject/onionoo/docs/BandwidthDocument.java
@@ -15,18 +15,24 @@ public class BandwidthDocument extends Document {
     this.fingerprint = fingerprint;
   }
 
-  @SuppressWarnings("unused")
   private Map<String, GraphHistory> write_history;
 
   public void setWriteHistory(Map<String, GraphHistory> writeHistory) {
     this.write_history = writeHistory;
   }
 
-  @SuppressWarnings("unused")
+  public Map<String, GraphHistory> getWriteHistory() {
+    return this.write_history;
+  }
+
   private Map<String, GraphHistory> read_history;
 
   public void setReadHistory(Map<String, GraphHistory> readHistory) {
     this.read_history = readHistory;
+  }
+
+  public Map<String, GraphHistory> getReadHistory() {
+    return this.read_history;
   }
 }